# backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Share Backend API",
    "description": "Share Backend API",
    "version": "v0.1.0"
  },
  "servers": [
    {
      "url": "https://hub.maayuan.fun:16666",
      "description": "Public HTTPS endpoint"
    }
  ],
...
  "paths": {
...
    "/v1/inventory/import": {
      "post": {
        "tags": [
          "库存"
        ],
        "summary": "导入库存/奖励",
        "operationId": "import",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryImportResult"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/open-api/inventory/import": {
      "post": {
        "tags": [
          "OpenAPI Inventory"
        ],
        "summary": "导入库存/奖励",
        "operationId": "inventoryImport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryImportResult"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiToken": []
          }
        ]
      }
    },
...
    "/v1/inventory/export": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "导出",
        "description": "导出(需登录;默认仅当前状态 full 快照,include=rewards 时附带区间奖励流水)。",
        "operationId": "export",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all"
              ]
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "current",
                "current,rewards"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryExportResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
    "/v1/inventory/current": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "当前库存",
        "description": "当前库存查询(需登录;entity_type 可选,缺省返回全部类型)。",
        "operationId": "current",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "item",
                "agent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListInventoryCurrentResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/open-api/inventory/export": {
      "get": {
        "tags": [
          "OpenAPI Inventory"
        ],
        "summary": "导出库存交换文档",
        "operationId": "inventoryExport",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "current",
                "current,rewards"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryExportResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiToken": []
          }
        ]
      }
    },
    "/open-api/inventory/current": {
      "get": {
        "tags": [
          "OpenAPI Inventory"
        ],
        "summary": "当前库存",
        "description": "当前库存(第三方只读示例,需 API Token 且具备 inventory:read 权限)",
        "operationId": "inventoryCurrent",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "item",
                "agent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListInventoryCurrentResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiToken": []
          }
        ]
      }
    },
    "/open-api/inventory/account": {
      "get": {
        "tags": [
          "OpenAPI Inventory"
        ],
        "summary": "Token 绑定的库存子账号",
        "operationId": "inventoryAccount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InventoryExchangeAccountDto": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          }
        },
        "required": [
          "id"
        ]
      },
...
      "InventoryImportRequest": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "myshare-inventory-exchange"
            ],
            "minLength": 1,
            "pattern": "^myshare-inventory-exchange$"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "maximum": 2,
            "minimum": 2
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerDto"
          },
          "catalog_version": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryExchangeAccountDto"
            },
            "maxItems": 1000,
            "minItems": 1
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryRecordRequest"
            },
            "maxItems": 1000,
            "minItems": 1
          }
        },
        "required": [
          "exported_at",
          "format",
          "producer",
          "records",
          "version"
        ]
      },
...
      "InventoryRecordRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/InventoryRewardRecord"
          },
          {
            "$ref": "#/components/schemas/InventoryFullSnapshotRecord"
          },
          {
            "$ref": "#/components/schemas/InventoryListedSnapshotRecord"
          }
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
          },
          "record_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          },
          "record_type": {
            "type": "string",
            "enum": [
              "reward_delta",
              "stock_snapshot"
            ],
            "minLength": 1,
            "pattern": "reward_delta|stock_snapshot"
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "item",
              "agent"
            ],
            "minLength": 1,
            "pattern": "item|agent"
          },
          "acquisition_channel": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "effective_at": {
            "type": "string",
            "format": "date-time"
          },
          "snapshot_scope": {
            "type": "string",
            "enum": [
              "full",
              "listed"
            ],
            "pattern": "full|listed"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryEntryRequest"
            }
          }
        },
        "required": [
          "account_id",
          "effective_at",
          "entity_type",
          "entries",
          "record_id",
          "record_type"
        ]
      },
...
      "ProducerDto": {
        "properties": {
          "platform": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$"
          },
          "version": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          }
        },
        "required": [
          "platform"
        ]
      },
      "InventoryError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "record_id": {
            "type": "string"
          },
          "entry_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "InventoryErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/InventoryError"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiResultInventoryImportResult": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/InventoryImportResult"
          }
        },
        "required": [
          "status_code"
        ]
      },
      "InventoryImportResult": {
        "properties": {
          "accepted": {
            "type": "integer",
            "format": "int32"
          },
          "duplicates": {
            "type": "integer",
            "format": "int32"
          },
          "history_only": {
            "type": "integer",
            "format": "int32"
          },
          "superseded": {
            "type": "integer",
            "format": "int32"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "accepted",
          "duplicates",
          "history_only",
          "superseded",
          "warnings"
        ]
      },
...
      "ApiResultInventoryAccountResponse": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/InventoryAccountResponse"
          }
        },
        "required": [
          "status_code"
        ]
      },
      "InventoryAccountResponse": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "created_at",
          "id",
          "name",
          "updated_at"
        ]
      },
...
      "InventoryExportAccountDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "InventoryExportRecordDto": {
        "properties": {
          "account_id": {
            "type": "string"
          },
          "record_id": {
            "type": "string"
          },
          "record_type": {
            "type": "string",
            "enum": [
              "reward_delta",
              "stock_snapshot"
            ]
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "item",
              "agent"
            ]
          },
          "acquisition_channel": {
            "type": "string"
          },
          "effective_at": {
            "type": "string",
            "format": "date-time"
          },
          "snapshot_scope": {
            "type": "string",
            "enum": [
              "full",
              "listed"
            ]
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryExportEntryDto"
            }
          }
        },
        "required": [
          "account_id",
          "effective_at",
          "entity_type",
          "entries",
          "record_id",
          "record_type"
        ]
      },
      "InventoryExportResponse": {
        "properties": {
          "format": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "default": 2,
            "enum": [
              2
            ],
            "maximum": 2,
            "minimum": 2
          },
          "exported_at": {
            "type": "string",
            "format": "date-time"
          },
          "catalog_version": {
            "type": "string"
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerDto"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryExportAccountDto"
            }
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryExportRecordDto"
            }
          }
        },
        "required": [
          "accounts",
          "exported_at",
          "format",
          "producer",
          "records",
          "version"
        ]
      },
      "ApiResultListInventoryCurrentResponse": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryCurrentResponse"
            }
          }
        },
        "required": [
          "status_code"
        ]
      },
      "InventoryCurrentResponse": {
        "properties": {
          "user_id": {
            "type": "string"
          },
          "account_id": {
            "type": "string"
          },
          "entity_type": {
            "type": "string"
          },
          "full_baseline_at": {
            "type": "string",
            "format": "date-time"
          },
          "entries": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StockEntryDto"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "account_id",
          "entity_type",
          "entries",
          "updated_at",
          "user_id"
        ]
      },
...
    },
    "securitySchemes": {
      "Jwt": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer"
      },
      "OpenApiToken": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API token"
      }
    }
  }
}